Bg model fix

+++ src/controllers/sales.js
@@ -1,23 +1,24 @@
-import Sales from "../model/Sales";
-
+import Sales from '../model/Sales';
+import Customers from '../model/Customers';
+import { v4 as uuidv4 } from 'uuid';
 class SalesController {
   static async getSales(req, res) {
     try {
       const foundSales = await Sales.find().populate({
-        path: "customerId",
-        model: "Customers",
+        path: 'customerId',
+        model: 'Customers',
       });
 
       if (foundSales.length === 0) {
-        return res.status(400).json({ message: "No Sales Found" });
+        return res.status(400).json({ message: 'No Sales Found' });
       }
 
       return res
         .status(200)
-        .json({ message: "Sales Found", Sales: foundSales });
+        .json({ message: 'Sales Found', Sales: foundSales });
     } catch (error) {
       return res.status(500).json({
-        message: "Server Error",
+        message: 'Server Error',
         Error: error.message,
       });
     }
@@ -26,30 +27,36 @@ class SalesController {
   static async getSale(req, res) {
     try {
       const { id } = req.params;
-      const sale = Sales.findById({ _id: id }).populate({
-        path: "customerId",
-        model: "Customers",
+      const sale = await Sales.findById({ _id: id }).populate({
+        path: 'customerId',
+        model: 'Customers',
       });
       if (!sale) {
-        return res.status(400).json({ message: "Sale not found" });
+        return res.status(400).json({ message: 'Sale not found' });
       }
 
-      return res.status(200).json({ message: "Sale Found", Sales: sale });
+      return res.status(200).json({ message: 'Sale Found', Sales: sale });
     } catch (error) {
       return res.status(500).json({
-        message: "Server Error",
+        message: 'Server Error',
         Error: error.message,
       });
     }
   }
 
   static async postSale(req, res) {
-    try {
-      const { orderNumber, customerName, products } = req.body;
+    const custId = req.params.customerId;
+    const customer = await Customers.findOne({ customerId: custId });
 
-      if (!orderNumber || !customerName || !products) {
+    if (!customer) {
+      return res.status(400).json('Not a registered customer');
+    }
+    try {
+      const { orderNumber, products } = req.body;
+
+      if (!orderNumber || !products) {
         return res.status(409).json({
-          message: "All fields are required",
+          message: 'All fields are required',
         });
       }
 
@@ -57,7 +64,7 @@ class SalesController {
         const { name, price, quantity } = product;
 
         if (!name || !price || !quantity) {
-          throw new Error("All fields are required");
+          throw new Error('All fields are required');
         }
 
         return {
@@ -72,44 +79,51 @@ class SalesController {
       }, 0);
 
       const sale = await Sales.create({
+        saleId: uuidv4(),
         orderNumber,
-        customerName,
+
+        customerId: custId,
         products: validProducts,
         totalAmount: validAmount,
       });
 
       return res.status(201).json({
-        message: "Sale Added",
+        message: 'Sale Added',
         sale: sale,
       });
     } catch (error) {
       return res.status(500).json({
-        message: "Server Error",
+        message: 'Server Error',
         error: error.message,
       });
     }
   }
 
   static async updateSale(req, res) {
+    const custId = req.params.customerId;
+    const customer = await Customers.findOne({ customerId: custId });
+
+    if (!customer) {
+      return res.status(400).json('Not a registered customer');
+    }
     try {
       const { id } = req.params;
 
       const sale = await Sales.findById(id);
 
       if (!sale) {
-        return res.status(400).json({ message: "Sale Not Found" });
+        return res.status(400).json({ message: 'Sale Not Found' });
       }
 
-      const { orderNumber, customerName, products, totalAmount } = req.body;
+      const { orderNumber, products, totalAmount } = req.body;
 
-      if (!orderNumber && !customerName && !products && !totalAmount) {
+      if (!orderNumber && !products && !totalAmount) {
         return res.status(409).json({
-          message: "No valid update data",
+          message: 'No valid update data',
         });
       }
 
       if (orderNumber) sale.orderNumber = orderNumber;
-      if (customerName) sale.customerName = customerName;
 
       if (products && Array.isArray(products)) {
         const validProducts = [];
@@ -120,7 +134,7 @@ class SalesController {
 
           if (!name || !price || !quantity) {
             return res.status(409).json({
-              message: "All fields are required",
+              message: 'All fields are required',
             });
           }
 
@@ -136,40 +150,45 @@ class SalesController {
         sale.products = validProducts;
         sale.totalAmount = validAmount;
       }
-
       const updatedSale = await sale.save();
 
       return res.status(200).json({
-        message: "Sale Updated",
+        message: 'Sale Updated',
         updatedSale,
       });
     } catch (error) {
       return res.status(500).json({
-        message: "Server Error",
+        message: 'Server Error',
         Error: error.message,
       });
     }
   }
 
   static async deleteSale(req, res) {
+    const custId = req.params.customerId;
+    const customer = await Customers.findOne({ customerId: custId });
+
+    if (!customer) {
+      return res.status(400).json('Not a registered customer');
+    }
     try {
       const { id } = req.params;
 
       const sale = Sales.findById({ _id: id });
 
       if (!sale) {
-        return res.status(400).json({ message: "Sale not found" });
+        return res.status(400).json({ message: 'Sale not found' });
       }
 
       const deletedSale = await Sales.findByIdAndDelete({ _id: id });
 
       return res.status(200).json({
-        message: "Sale Deleted",
+        message: 'Sale Deleted',
         deletedSale,
       });
     } catch (error) {
       return res.status(500).json({
-        message: "Server Error",
+        message: 'Server Error',
         Error: error.message,
       });
     }
